pyproject.toml
[tool.poetry]
name = "codeleap-test"
version = "0.1.0"
description = "Test project for Codeleap, made with Django and DRF"
authors = ["Spearks <otavio.canuto.m@gmail.com>"]
license = "AGPL V3"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.1"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
